+++ liana-gui/src/app/view/label.rs
@@ -12,7 +12,7 @@ use crate::app::view;
 pub fn label_editable(
     labelled: Vec<String>,
     label: Option<&String>,
-    size: u16,
+    size: u32,
 ) -> Element<'_, view::Message> {
     if let Some(label) = label {
         if !label.is_empty() {
@@ -46,7 +46,7 @@ pub fn label_editable(
 pub fn label_editing(
     labelled: Vec<String>,
     label: &form::Value<String>,
-    size: u16,
+    size: u32,
 ) -> Element<'_, view::Message> {
     let e: Element<view::LabelMessage> = Container::new(
         row!(
@@ -71,7 +71,7 @@ pub fn label_editing(
 pub fn label_non_editable(
     labelled: Vec<String>,
     label: Option<&String>,
-    size: u16,
+    size: u32,
 ) -> Element<'_, view::Message> {
     let label_text = label.map(|s| s.as_str()).unwrap_or("(External Output)");
